Added motivation bar to main scene; Added accessor for Harvest Crops minigame; Added floatRatio to motivation manager
Motivation meter wouldn't work properly due to its 0-1 float constraint.  Added a float ratio in the manager script to fix that.

+++ Assets/Scripts/UI/Main/MotivationMeter.cs
@@ -19,7 +19,7 @@ public class MotivationMeter : MonoBehaviour
         MotivationManager.OnMotivationUpdated -= OnMotivationUpdated;
     }
 
-    private void OnMotivationUpdated(int newVal)
+    private void OnMotivationUpdated(float newVal)
     {
         if (fillImage == null) return;
